[refac] 폴더 구조 좀더 자세히 설계 변경
