+++ be/src/services/tweet/getTweet.ts
@@ -1,4 +1,4 @@
-import { userModel, tweetModel } from '../models';
+import { userModel, tweetModel } from '../../models';
 
 const getFollowingTweetList = async (_: any, args: any) => {
   const userId = args.id;
@@ -69,46 +69,4 @@ const getUserTweetList = async (_: any, args: any) => {
   return tweetList;
 };
 
-const addBasicTweet = async (_: any, args: any) => {
-  const userId = 'test1';
-  const content = '트윗 내용';
-  const imgUrls = [] as Array<String>;
-  const newTweet = await tweetModel.create({
-    author_id: userId,
-    content: content,
-    img_url_list: imgUrls,
-    child_tweet_list: [],
-  });
-  console.log(newTweet);
-  return newTweet;
-};
-
-const addReplyTweet = async (_: any, args: any) => {
-  const userId = 'test1';
-  const content = '트윗 내용';
-  const imgUrls = [] as Array<String>;
-  const parentId = '5fbca83023d2695f10ab52b4';
-
-  const replyTweet = await tweetModel.create({
-    author_id: userId,
-    content: content,
-    img_url_list: imgUrls,
-    parent_id: parentId,
-    child_tweet_list: [],
-  });
-  console.log(replyTweet);
-
-  const childId = replyTweet?.get('_id');
-
-  const parentTweet = await tweetModel.findOneAndUpdate(
-    { _id: parentId },
-    { $push: { child_tweet_list: childId } },
-    { new: true },
-  );
-
-  console.log(parentTweet);
-
-  return replyTweet;
-};
-
-export { getFollowingTweetList, getUserTweetList, addBasicTweet, addReplyTweet };
+export { getFollowingTweetList, getUserTweetList };
