+++ be/src/services/user/addUser.ts
@@ -1,0 +1,1 @@
+import { userModel } from '../../models';
