+++ be/src/services/user/getUser.ts
@@ -1,4 +1,4 @@
-import { userModel } from '../models';
+import { userModel } from '../../models';
 
 const getFollowingList = async (_: any, args: any) => {
   const userId = args.id;
@@ -40,26 +40,4 @@ const getSearchedUserList = async (_: any, args: any) => {
   return followerList;
 };
 
-const followUser = async (_: any, args: any) => {
-  const userId = 'test3';
-  const followUserId = args.follow_user_id;
-  const user = await userModel.findOneAndUpdate(
-    { user_id: userId },
-    { $addToSet: { following_list: followUserId } },
-    { new: true },
-  );
-  return user;
-};
-
-const unfollowUser = async (_: any, args: any) => {
-  const userId = 'test3';
-  const unfollowUserId = args.unfollow_user_id;
-  const user = await userModel.findOneAndUpdate(
-    { user_id: userId },
-    { $pull: { following_list: unfollowUserId } },
-    { new: true },
-  );
-  return user;
-};
-
-export { getFollowerList, getFollowingList, getSearchedUserList, followUser, unfollowUser };
+export { getFollowerList, getFollowingList, getSearchedUserList };
