+++ be/src/services/user/index.ts
@@ -1,0 +1,4 @@
+import { getFollowerList, getFollowingList, getSearchedUserList } from './getUser';
+import { followUser, unfollowUser } from './modifyUser';
+
+export { getFollowerList, getFollowingList, getSearchedUserList, followUser, unfollowUser };
